+++ priv/repo/migrations/20240307141604_create_prompts.exs
@@ -5,7 +5,7 @@ defmodule WtrSurvey.Repo.Migrations.CreatePrompts do
     create table(:prompts) do
       add :title, :string
       add :participant, :string
-      add :max, :float
+      add :max, :integer
       add :factor, :float
 
       add :survey_id, references(:surveys, on_delete: :delete_all), null: false
